Fix how backend secrets are handled
When the .env file is not explicitly stated and the docker-compose file has
environment variables declared, those values seem to supersede any secrets set
in the .env file. This was causing stupid errors (like being unable to reach
the backend due to a gateway error or Google claiming that no client ID was
found), so I was finally forced to fix it.