Add game state db table
I'm not really excited about a document-like text field in a relational
database, but I also don't want to keep spending time trying to iterate on the
perfect schema. When I've cemented the game play and have a better idea of what
an appropriate schema looks like, I can upgrade this.